refactor(pages/home): extract `<Header />`
