+++ utils/classNames.ts
@@ -1,0 +1,3 @@
+export function classNames(...classes: (string | undefined)[]) {
+  return classes.filter(Boolean).join(" ");
+}
